Create Rand_Num_Char_String.ino
Can you write a random number+alphanumeric string generator for an Adafruit Metro M4 express board (ATSAMD51 chipset)? 

The Adafruit Metro M4 express board uses the ATSAMD51 chipset, which has a built-in hardware random number generator (RNG), so we can take advantage of that to generate random numbers. However, the Arduino random() function may not use the hardware RNG, and you need to use the RAND register of the ATSAMD51 directly to get random numbers from the hardware RNG.

In order to convert the generated random numbers to alphanumeric strings, we will use the same approach as in the previous Arduino Uno example.

Here's an example sketch that generates random numbers and alphanumeric strings:

This sketch first initializes the hardware RNG in the setup() function. Then, in the loop() function, it generates a random number, converts it to an alphanumeric string, and prints both the number and the string to the Serial Monitor. Each random number and corresponding alphanumeric string is printed once per second.