Create RandNum_Str_Pseudowrd.ino
Refactor the previous code by adding the ability to choose the size of the random numbers and Alphanumeric strings. Add a pseudoword generator to create pronounceable, but not necessarily meaningful, English-like words. These pseudowords should output in addition to the random numbers and alphanumeric strings.

In order to generate pseudowords, we'll need a library or algorithm that can construct pronounceable sequences of letters. There isn't a straightforward way to do this, but we can improvise by using a vowel and consonant alternation pattern for simplicity.

Here's the refactored sketch:

In this version, the size of the alphanumeric string and pseudoword can be set by changing the length variable in the loop() function. The randAlphanumeric() function takes a random number and a length as inputs and generates an alphanumeric string of the specified length. The generatePseudoWord() function generates a pseudoword of the specified length by alternating between randomly selected consonants and vowels.