Create MetroM4_FFTBenchmarking.ino
This sketch generates random test data in the setup() function and then in the loop() function, it runs the FFT and inverse FFT for different sizes, from 16 to 1024. For each size, it prints the FFT size, the FFT results, and the inverse FFT results.

Please note that the inverse FFT results will be somewhat distorted due to the windowing function and the finite precision of the FFT and inverse FFT, but they should resemble the original data. This is an inherent limitation of the FFT and inverse FFT algorithms and is not specific to the ArduinoFFT library.