Create RandPseudwrdGen_MetroM4_Simple.ino
Generate a random alphanumeric string generator code for the Adafruit Metro M4 express board (ATSAMD51 chipset). Be sure to include the ability for users to choose the size of the random numbers and Alphanumeric strings in the code by the user. The alphanumeric strings should include capital and lower case letters as well as numbers and other special characters. Add a pseudoword generator to create pronounceable, but not necessarily meaningful, English-like words. All numbers and strings so generated should be labeled and output visible from the serial monitor and plotter. Finally suggest a file name for this code.

This code does not use any non-standard libraries and should compile correctly in your Arduino IDE. The code generates random numbers, random alphanumeric strings, and pseudowords, and outputs the results to the serial monitor.