Create CustRanNumPsudwordGen_MetroM4_Simple.ino
Refactor that previous code and modify the size of the random numbers so that they vary from 1-999999 when generated. Also modify the size of the random alphanumeric strings and pseduowords. Make sure to provide the option for users to modify the code to output specific output sizes/string lengths in the code.

In this code, users can change the NUMBER_MIN, NUMBER_MAX, STRING_LENGTH, and PSEUDOWORD_LENGTH to suit their preferences. The code will then generate random numbers, random alphanumeric strings, and pseudowords of the specified lengths, and output the results to the serial monitor.