Create HiddenMarkovModel_WeatherPrediction.ino
Implementing a Hidden Markov Model (HMM) on an ATSAMD51 chipset such as the Adafruit Metro M4 board can be quite complex, particularly because an HMM typically involves substantial mathematical computation and data handling that can be challenging on microcontroller environments.

However, it's possible to implement a simplified version. As an example, let's consider a weather prediction system, where we try to predict if it will be rainy or sunny based on the previous day's weather.

Firstly, you need to define the states and the transition probabilities between states. For instance, we might define our states as "Rainy" and "Sunny". We could then define our transition probabilities - the chance of moving from one state to another - based on historical data or assumptions.

For the sake of this example, we could have:

The chance of transitioning from "Sunny" to "Rainy" as 0.1 (10% chance of rain after a sunny day).
The chance of transitioning from "Rainy" to "Sunny" as 0.3 (30% chance of a sunny day after a rainy one).
The chance of staying in the "Sunny" state (i.e., having a sunny day after a sunny day) as 0.9.
The chance of staying in the "Rainy" state as 0.7.
Here is a simple implementation of a weather prediction system using these probabilities:

This simple code does not fully utilize the power of HMMs, but it provides a basic illustration of the principle.

Real HMMs often involve more complex computations including forward-backward algorithms, the Viterbi algorithm for finding the most likely sequence of states, and the Baum-Welch algorithm for training the HMM. These are typically too computationally intense to be run on a microcontroller and are better suited to more powerful environments like a PC, a server, or cloud-based solutions.

As such, if you need to use HMMs in a microcontroller setting, it may be best to perform the complex computations elsewhere and then upload the computed transition probabilities to the microcontroller for it to use.