Create MonteCarloIntegration_PiEstimation.ino
How can we implement a Monte Carlo integration demo program for the ATSAMD51 chipset Adafruit Metro M4 board? 

Monte Carlo integration is a statistical method used to compute the value of an integral. It's based on the law of large numbers and uses random sampling to obtain numerical results.

The idea is simple: if you want to compute the integral of a function over a given interval, you can approximate this integral by taking random samples from the interval and averaging the function values at these samples.

Here's a simple example of how you can implement Monte Carlo integration on the ATSAMD51 Adafruit Metro M4 board to calculate the value of pi:

In this demo, we randomly generate points in a 1x1 square and count how many of these points fall within a circle of radius 1 inscribed in this square. The ratio of the points inside the circle to the total number of points is approximately pi/4. Therefore, we can estimate the value of pi by multiplying this ratio by 4.

Suggested filename: MonteCarloIntegration_PiEstimation.ino

That's a good code but can we refactor it in order to take advantage of the ATSAMD51 hardware random number generator capabilities?

In this updated code, we're using the rand_sync_read32 function from the hri_rand_sync.h library to generate random numbers using the HRNG of the ATSAMD51.

Note that to use the HRNG, you will need to have the appropriate Atmel START libraries installed. You can add these libraries to your Arduino environment by creating a custom Atmel START project for the ATSAMD51 and exporting it as an Arduino library.

Can we refactor the code without having to install any libraries beyond the standard arduino IDE libraries?

Yes, that's possible, however, without additional libraries, it might not be possible to access the true hardware random number generator of the ATSAMD51. Instead, we can use the built-in random() function in Arduino, which is a pseudo-random number generator.

Here's the refactored code:

This code will still give you a reasonable approximation of pi, but it's worth noting that the random() function in Arduino isn't truly random, it's a pseudo-random number generator. This means it uses a deterministic algorithm to generate a sequence of numbers that appears random. The randomSeed() function is used to initialize this algorithm with a somewhat random input (in this case, the value of an unconnected analog pin).

The suggested filename is: MonteCarloIntegration_PiEstimation.ino