Create SAMD51_TRNG_Demo.ino
Description: "This program demonstrates the functionality of the True Random Number Generator (TRNG) on the SAMD51 microcontroller. It continuously generates a series of random numbers using the built-in TRNG and outputs them to the serial monitor."

It appears the SAMD51 does not have TRNG turned on by default so this file demonstrates what turning it on looks like (verified Metro M4 board). 

You need to open the serial monitor and input keystrokes which appear to supply the entropy to generate a few random numbers. ***No author attribution found on Arduino Forums***