Create RTC_UserSet_SAMD51.ino
RTC clock on a SAMD51 microcontroller can be set by user input via the Serial Monitor using the RTC_SAMD51.h library(https://github.com/Seeed-Studio/Seeed_Arduino_RTC/blob/master/examples/rtc_samd51/rtc_samd51.ino). This library allows you to adjust the time, print the current time, the day of the week, time since midnight, and calculate a future date/time.

Here's a refactored version of the code:

This code will prompt the user to enter the current date and time in the format YYYY/MM/DD HH:MM:SS, parse that input, and then adjust the RTC to the user-provided date and time. The current date and time (including the day of the week) will be printed to the Serial Monitor in the main loop.