Create RandStrngPsudwrdGen_Metro4KolgComp.ino
Computing the exact Kolmogorov complexity of a string is an uncomputable problem, meaning that there is no algorithm that can determine it for all strings. However, for practical purposes, we can estimate it by using the length of the compressed string. Here's a refactored version of your code, that uses a simple compression algorithm (Run-Length Encoding) to estimate the Kolmogorov complexity. Note that this estimate is very rough and there are better compression algorithms that could provide a better estimate:

This refactored code adds an additional function called estimateKolmogorovComplexity to provide an estimate of the Kolmogorov complexity using run-length encoding for compression. It then prints the estimated Kolmogorov complexity for each string generated.