Create TRNG_ChainRuleKolmogorovComplexity.ino
The Chain Rule for Kolmogorov complexity states that for two strings x and y, the complexity of the concatenation of the strings x + y can be approximated as K(x + y) ≈ K(x) + K(y|x), where K(y|x) is the complexity of y conditioned on x. However, it is important to note that Kolmogorov complexity is not computable. There is no algorithm that can take a string as input and output its Kolmogorov complexity. Instead, I'll use a simple approximation of the chain rule by considering the compression of the strings as we did before. This won't be a real Kolmogorov complexity but can serve as a rough measure.

Below is the refactored code:

Please note that this code is using a very basic approximation technique and doesn't actually compute the true Kolmogorov complexity, which is an uncomputable property. This is more of a teaching example to understand the concept of Kolmogorov complexity and the chain rule.