Create Slave_TRANSEC_Key_Exchange.ino
Generate the code for exchange of TRANSEC keys over SPI. Suggest a file name upon completion.

For exchanging TRANSEC keys over SPI, you would need two separate pieces of code, one for the master device and another for the slave device. The master initiates communication and sends the TRANSEC key to the slave.

Here is the code for the Slave Device.

This is a basic example of how the keys can be exchanged between two devices. Note that for any real-world application, security measures such as encryption should be used.

Suggested file names: Master_TRANSEC_Key_Exchange.ino for the master device and Slave_TRANSEC_Key_Exchange.ino for the slave device.