Create RandomCodeWordsGenerator.ino
Generate a Random Code Words program for the SAMD51 using the TRNG in C++ for use on the Arduino IDE. Suggest a filename upon completion. Remember that the SAMD51 TRNG must be initialized via the following code example.

// Function to initialize the hardware random number generator on the ATSAMD51
void initRNG() {
  // Enable the APBC clock for the RNG
  MCLK->APBCMASK.reg |= MCLK_APBCMASK_TRNG;

  // Set up and enable the RNG
  TRNG->CTRLA.reg = TRNG_CTRLA_ENABLE;
}

// Function to generate a random number using the hardware RNG
uint32_t getRandomNumber() {
  while (TRNG->INTFLAG.bit.DATARDY == 0);  // Wait for a random number to be ready
  return TRNG->DATA.reg;  // Read the random number
}

void setup() {
  Serial.begin(9600);
  while (!Serial);

  initRNG();
}

void loop() {
  uint32_t randNum = getRandomNumber();
  char alphanumeric[9];
  
  for(int i = 0; i < 8; i++) {
    uint8_t randVal = (randNum >> (i * 4)) & 0x0F;
    if (randVal < 10)
      alphanumeric[i] = '0' + randVal;
    else
      alphanumeric[i] = 'A' + (randVal - 10);
  }

  alphanumeric[8] = '\0';  // Null-terminate the string

  Serial.print("Random Number: ");
  Serial.print(randNum);
  Serial.print(", Alphanumeric String: ");
  Serial.println(alphanumeric);

  delay(1000);
}

The code you provided is almost a complete example for generating random code words using the SAMD51 TRNG. Below is a slightly modified version that focuses on generating random alphanumeric code words. It initializes the TRNG, generates random numbers, and converts them into alphanumeric strings that can be used as code words.

This program initializes the TRNG of the SAMD51, generates random alphanumeric characters, and combines them to form code words. It prints a new random code word to the serial port every second. Note that the randomCharacter function uses the getRandomNumber function to generate a random index for selecting a character from the alphanumeric range ('0' to '9' and 'A' to 'Z').