Create SeededRandomCodeWordsGeneratorWithInfo.ino
The following code refactors the previous example to include a serial monitor prompt asking the user to enter an alphanumeric string. The user input is used as an additional source of entropy by XORing each character of the input with the least significant byte of the random numbers generated by the TRNG.

This program initializes the TRNG and waits for the user to input an alphanumeric string via the Serial Monitor. Once the user presses enter, it uses the entered string to seed the random number generation. The user input is XORed with the random numbers generated by the TRNG to introduce an additional element of randomness. It prints a new random code word to the serial port every second.

In the modified code below, the user is informed about the parameters of the desired string that should be entered in the serial monitor prompt. The code is also modified to accommodate long strings of random alphanumeric keystrokes.

This version of the program gives the user more information on the expected input and allows for long strings of alphanumeric characters. If the user enters a character that is not alphanumeric, an error message is displayed. The input string is used to seed the random number generation, as before. The program continues to generate random code words and outputs them to the serial monitor.

NOTE: In the Arduino Serial Monitor window, you will find an option at the bottom right corner to select how the line endings are sent when you press Enter. You need to make sure that it is set to either 'Newline' or 'Both NL & CR', because the code is waiting for a newline character '\n' to know when the user has finished entering the string.