Create SeededRandomCodeWordsGenwHash.ino
The inputSeed variable was not being used properly to seed the random number generator. Instead of just using a character from the input string, it's better to hash the entire string and use that as a seed. Here's an updated version of the code where I use a simplified hashing function to combine all characters in the input string.

 In the Arduino Serial Monitor window, you will find an option at the bottom right corner to select how the line endings are sent when you press Enter. You need to make sure that it is set to either 'Newline' or 'Both NL & CR', because the code is waiting for a newline character '\n' to know when the user has finished entering the string.

Please also make sure that in the Serial Monitor window, the line ending option is set to 'Newline' or 'Both NL & CR'. This will send the newline character '\n' when you press Enter, and the program will recognize that as the end of the input string.